pyproject.toml
[project]
name = "langgraph-agent"
version = "0.1.0"
description = "Minimal LangGraph agent with CLI"
authors = [{ name = "You" }]
requires-python = ">=3.10"
dependencies = [
	"langgraph>=0.2.30",
	"langchain-core>=0.2.43",
	"langchain-openai>=0.1.20",
	"pydantic>=2.8",
	"python-dotenv>=1.0.1",
	"rich>=13.7.1",
	"typer>=0.12.3",
	"openai>=1.35.0",
]

[project.scripts]
agent = "agent.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"